chore: refactor feedback updating and action handling
